ai-project/실습/ex02/requirements-optional.txt
python-docx
pillow
pytesseract
openpyxl
fastmcp
